Version 1.0.3 - Fixed BGG links with gg-markup-safe-html
Some BGG <a> links have children that are <gg-markup-safe-html> elements. The extension was failing on these since when you hover over the <a> you initially hover over the <a> element, but a millisecond later you are hovering over the <gg-markup-safe-html> element. To fix this, when deciding whether the mouse has left the <a> element, we double-check if the mouse is now hovering over a child of the <a> element as well, and do not call onLeave if we are hovering over a child element.